Move long examples to their own files
